Increase resolution of the TFL DAR logo
We’re showing these kinds of logos bigger as of this PR:
https://github.com/alphagov/notifications-utils/pull/170

If we don’t increase the resolution of the asset to be bigger then it
will look blurry.

New image is `108px` high, which is `2 * 54px`, in order for it to look
as good as possible on retina screens.