Test for page titles in client request fixture
We’re frequently ending up with mismatched page titles on our pages
because:
- they’re hard to spot in the browsers tab bar
- new pages are often created by copy pasting old ones

It would be hard to write some kind of macro that spat out the h1
and the page title because they go into different blocks in the
parent template.

Instead, we can catch these mistakes by testing for them. Going
forward we should be using the `client_request` fixture for testing.
So it makes sense to put the page title test in here so that we
get it for free in every test – same way we do for testing response
codes.

fixup! Test for page titles in client request fixture