Add some internal navigation for platform admin
There are way too many links in the top nav now (when you’re using the
site as a platform admin).

And how often are you going to look at the letter jobs page really?

This commit adds some internal navigation for the platform admin area of
the platform, copying the style that works for navigating a single
service.