Remove stuff about test jobs on row errors page
The send yourself a test flow doesn’t reference any kind of
file/rows/table in the front end any more, and doesn’t create a CSV in
the background (except for letters). And it should validate any mistakes
before getting to the check page.

So any user doing a send yourself a test flow should never get to this
template, which means we can remove some redundant code.