Truncate inbound message content
Inbound text messages can run over multiple lines. This makes the page
harder to scan. Your phone, and the outbound messages page, only show
the first line of the text message, and truncate the rest with an
ellipsis.

This commit does the same for inbound text messages.

It also stops the timestamp for the inbound messages being squashed and
wrapping over multiple lines, which looks messy.

We couldn’t do this before, because it would have stopped people from
being able to copy/paste the full message content from this page.