Standardise Google Analytics tracking across GaaP
These are the settings that our analytics person has said we should be
using across all the GaaP products.

This commit also makes sure our tracking code is identical across all
the templates that have it in (including the obsfucation of UUIDs). We
may want to remove the ID obsfucation later on, but for now let’s make
sure it’s happening consistently in all the places.