Strip whitespace before validating any form
We’ve had whitespace-in-emails problems on:
- register form
- sign in form
- send one off form

We should just handle whitespace gracefully, by ignoring it. Makes sense
to do this at a global level because there might be other places it’s
causing problems that we don’t know about.

The only place that users _might_ be relying on leading/trailing spaces
(and we have no way of knowing this) is in passwords or tokens. So this
filter ignores any instances of `PasswordField`.

Adapted from @pcraig3’s work on Digital Marketplace:
- https://github.com/alphagov/digitalmarketplace-supplier-frontend/commit/4b06f5cfb703678f5b8bb303e2f0590bb248b5f2
- https://github.com/alphagov/digitalmarketplace-supplier-frontend/commit/e761c1ce655137fed11f730da4e0a2b1893dee91