Fix bug with disappearing AJAX content in IE 9/10
We had a user report that the ‘inbound messages’ part of their dashboard
was appearing for a split second then disappearing.

This was caused by the Javascript on the page throwing an exception as
it was trying to re-render this part of the page. This meant it gave up
and rendered nothing.

The exception was caused by passing `undefined` as the second argument
to `.insertBefore`. This is acceptable in most browsers, but not older
versions of IE. This is fixed in the latest version of diffDOM (the 3rd
party library we use to do the AJAX stuff) by defaulting the second
argument to `.insertBefore` to `null`, which is acceptable in old IE
versions.

See the fix here: https://github.com/fiduswriter/diffDOM/commit/8833d87e9d77d03ab94e792875697ab00712a714